ocr_service/requirements.txt
# PyTorch ROCm — installed separately by setup.sh via --index-url
# torch, torchvision, torchaudio

# Document processing
docling>=2.7.0
onnxruntime>=1.17.0
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.0

# API framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0

# Utilities
psutil>=5.9.0
pytest>=8.0.0
httpx>=0.27.0
fpdf2>=2.7.0
